Moved JournalCollector to other package to encourage reuse
